requirements.txt
pandas==1.3.1
streamlit==1.12.2
plotly_express==0.4.0


